Overhaul the save mechanism
Previously we would load all vehicles at once. Now we have a harder time
knowing when is a good time for that. So every vehicle requests reload
for itself. Unfortunately, it also requests reload for every other
enrolled vehicle, but whatever. It's a small loss, and save/load is
functionally correct.