probable speedup
This update concerns using a fabricator in a ModVehicle. It should spend the ModVehicle's energy. This logic isn't very keen; we check every MV to see if the player is in one. If so, that's the one whose energy we spend. This update adds the opportunity for the object to directly find the MV by looking up and down its transform hierarchy.