Issue a warning about broken ping instances
One time I had a Constructor that, upon loading into the game, was missing its ping instance and got issued one. But the new ping instance didn't get an origin, so it would throw an error on HUD ping update. That caused every HUD ping after the constructor to not be shown. I had to pick up the constructor to fix it. This is just a helpful log line to identify what causes the issue.