Remedy color picker issues
Sometimes the color picker gets disabled in one of two ways: the canvas is disabled, or the active screen is disabled. I don't know why this happens. This update makes it so that VF ensures the color picker is active (in both these ways) every time the player enters the vehicle. I wish I knew what caused this issue!