Improve moongate handling
I was previously unaware of the moongate at the lava castle. That one, for some reason, didn't work with the BlockSeamoth patch. So it's been improved. In addition, the precursorUnderwater field is now tracked appropriately, so even if you manage to get inside and past the BlockSeamoth collider, you'll be "out of water" and helpless