Fix magnet boots saveload issues
It seems a booted MV does not properly receive the save and load signals from IProtoTreeEventListener. So we make sure that before save happens, magnet boots are detached and the user is notified when this happens. Would love to reattach them, but would have to smartly wait for every json action to be finished before doing so.