Refactor deep water patch (easycraft)
EasyCraft also has a prefix patch on this method. It always returns false and replaces the whole method. For that reason, "craft" would get called twice when EasyCraft was present. Change it to a postfix in order to check if the constructor is active before trying anything. In the event this code changes the behavior of the constructor, remove the voice line that said it wouldn't work.