Fix save/load issues for MagnetBoots
In some cases, the previous implementation would fail. For example, a seamoth attached to a cyclops. In this case, the cyclops would try to dock the seamoth internally, which would make the seamoth inaccessible. The new implementation seems to be okay in all ways, except that cyclops-attached seamoths are saved as being "not attached," so on reload, they will not attach to their prior parent cyclops.